Make SPICE2 compile and work on 64-bit system
SPICE2 has some fairly obnoxious "memory management" that looks like
it is intended to be portable between systems with different word
lengths, but in fact only works if integers are 4 bytes, double
precisions are 8, and complex are 16, and pointers are 4 bytes (32 bit
system).

This commit hamfistedly hacks in changes that allow SPICE2 to build
and run correctly on 64-bit systems, at the expense of breaking the
ability to build on 32-bit.

We have to change several things:
  - The unix.c port file must define the "loc_" function so it can
    return 64-bit pointers instead of 32-bit.  That requires long int
    return value.
  - The itoc function needed its char array to be static.
  - Because SPICE2 freely uses pointers returned by LOC as if they are
    integers, integers must be 8 bytes long (long ints), which
    requires using gfortran's "-finteger-4-integer-8" option (forcing
    all integers in the program to 8 bytes
  - The LOCF function returns a "pointer" that is in units of the
    integer size, not in bytes.  So we must divide the pointer
    returned by 8 instead of 4 (see previous point).
  - Finally, SPICE2 has paid lip service to being portable to
    different word sizes (see how they try to use NWD4, NWD8, NWD16 as
    the size (in integers) of integers, doubles, and complex), but
    in fact are EXPLICITLY assuming that integers are 4 bytes, doubles
    8, and complex 16 by their use of the "zero4," "zero8," and
    "zero16" functions to zero out blocks of integers, doubles, and
    complex, respectively.  The unix.c definitions of these functions
    dutifully zero out 4, 8, and 16 byte blocks as implied by their
    names.  The zero4 is now wrong for integers, because they're
    actually 8.  So let's fake it out by making zero4 actually zero 8
    byte blocks under the hood.

This commit, as I said, makes spice2 compile, link, and PRODUCE
CORRECT RESULTS on a 64 bit system.

  ./spice < benchmark.in > benchmark.test

results in a benchmark.test file that differs from benchmark.out ONLY
in the dates and in the formatting of double precision numbers ("E" or
"D" instead of "e" in exponents, mostly).

It does, however, completely break it for 32-bit, so is not the best
solution.  More careful work would need to be done to make this truly portable
between 32 and 64 bit systems.

Finally, I have modified "build" so that it does not use the linuxism
of assuming that bash lives in /bin.  The script isn't even
bash-specific, so use the completely portable "/bin/sh" which exists
on all Unix-like systems, not just Linux.

Note that I have explicitly used "-O0" in this compilation.  I have
found that using "-O3" inexplicably causes the resulting binary to get
parse errors on the "benchmark.in" file, whereas the unoptimized build
runs that input just fine.

https://github.com/ethanmancini/SPICE-2/issues/1