Merge mac-os branch and clean up makefile
For some reason, a bunch of commits on the mac-os branch were never
merged to master, but were useful everywhere so should have been.

This commit merges those changes into master, and then cleans up the
makefile so it references CC and CFLAGS (C compiler and its flags)
instead of using the CXX and CXXFLAGS variables that are usually only
used for C++.

I also added the "-Wno-error" flag to GCC's flags, because unix.c has
a handful of functions that would have been declared "void" in modern
C because they return no value, but instead were declared with no
type.  GCC has always warned of this usage (which is harmless in this
case) but apparently some newer systems have "-Werror" as default and
make these warnings show up as fatal compilation errors.  "-Wno-error"
turns that behavior off and lets them just be warnings.