Apply changes to remove meaningless warnings
The code in unix.c defined many functions meant to be called from
Fortran that returned no value, or that were meant to be called
internally but also returned no value and were used without trying to
access any return value.

This meant that all the undeclared functions were presumed to be of
type int, which is the old fashioned assumption in early C, and causes
no problems in this case.  It is, however, warned of by all modern
compilers because it is often a mistake to do that.

Recent versions of gcc have begun treating such warnings as errors and
defining that type of usage as "nonconformant" (that is, not in
conformance with modern coding best practices).

GH-6 reports compilation failures due to these issues, and the
workaround was to use compiler options to turn those particular issues
back into warnings.  GH-8 is a pull request that puts those options
into Makefile, which is a fine approach, but a better one is to make
the unix.c code not throw warnings in the first place.

This commit fixes the code so it no longer throws warnings.

All functions meant to be called from fortran as subroutines (those
that return no value and whose names end with "_") have been declared
as void.

All functions called internally only by unix.c and which return no
value have been declared as void, and prototypes added to the top of
the file to prevent warnings about redefining function types that had
previously been declared implicitly as int.

Includes of stdlib.h and sys/times.h have been inserted to prevent
warnings about implicit declaration of standard system functions
"exit" and "times" (respectively).

Correct casting of an integer pointer to a (struct tms *) pointer has
been inserted to prevent warnings about incorrect argument type to the
"times" system function.

This commit should render GH-8 unnecessary, and should fix GH-6.

Closes GH-6