Fix warnings issued by clang
Clang is throwing even more errors than gcc was, reporting usages that
will be illegal in the C23 standard.

Most of these are related to this ancient code using K&R style
function declarations, such as:

```
   void function(a,b)
   int a,b;
```
instead of modern ANSI C's
```
   void function(int a,int b)
```

Some are also due to this file having previously failed to include
sys/times.h, and instead just declaring every one of the standard
functions defined there directly, and in a manner that is now
incompatible with the declarations in sys/times.h that is now
included, and which would be illegal in C23.

This commit fixes all of those, and now unix.c compiles without error
or warning in *either* GCC or Clang.